+++ js/imageeditor/Index.svelte
@@ -29,7 +29,7 @@
 	export let root: string;
 	export let value_is_output = false;
 
-	export let height: number | undefined = 400;
+	export let height: number | undefined = 450;
 	export let width: number | undefined;
 
 	export let _selectable = false;
@@ -91,7 +91,9 @@
 
 	$: value && handle_change();
 	const is_browser = typeof window !== "undefined";
-	const raf = is_browser ? window.requestAnimationFrame : () => {};
+	const raf = is_browser
+		? window.requestAnimationFrame
+		: (cb: (...args: any[]) => void) => cb();
 
 	function wait_for_next_frame(): Promise<void> {
 		return new Promise((resolve) => {
@@ -216,8 +218,8 @@
 			accept_blobs={server.accept_blobs}
 			{layers}
 			status={loading_status?.status}
-			upload={gradio.client.upload}
-			stream_handler={gradio.client.stream}
+			upload={(...args) => gradio.client.upload(...args)}
+			stream_handler={(...args) => gradio.client.stream(...args)}
 			{placeholder}
 		></InteractiveImageEditor>
 	</Block>