+++ js/spa/test/state_change.spec.ts
@@ -56,3 +56,30 @@ test("test datastructure-based state changes", async ({ page }) => {
 	await expect(page.getByLabel("Changes")).toHaveValue("2");
 	await expect(page.getByLabel("Clicks")).toHaveValue("5");
 });
+
+test("test generators properly trigger state changes", async ({ page }) => {
+	await page.getByRole("button", { name: "Iterator State Change" }).click();
+	await expect(page.getByTestId("markdown").first()).toHaveText(
+		"Success Box 0 added"
+	);
+	await page.getByRole("button", { name: "Iterator State Change" }).click();
+	await expect(page.getByTestId("markdown").nth(1)).toHaveText(
+		"Success Box 1 added"
+	);
+});
+
+test("test state change for custom hashes", async ({ page }) => {
+	await expect(page.getByLabel("Custom State Changes").first()).toHaveValue(
+		"0"
+	);
+	await page.getByRole("button", { name: "Set State to 10" }).click();
+	await expect(page.getByLabel("Custom State Clicks").first()).toHaveValue("1");
+	await expect(page.getByLabel("Custom State Changes").first()).toHaveValue(
+		"1"
+	);
+	await page.getByRole("button", { name: "Set State to 10" }).click();
+	await expect(page.getByLabel("Custom State Clicks").first()).toHaveValue("2");
+	await expect(page.getByLabel("Custom State Changes").first()).toHaveValue(
+		"1"
+	);
+});